chore: update test-runner to match `master` functionality
Cherry-picked changes from 22122c4a5833ae961c04a9ff2571b9d43da428dc, but
only for `test-runner.js`.

+++ e2e/test-runner.js
@@ -4,6 +4,11 @@ const fs = require('fs-extra')
 const chalk = require('chalk')
 
 const IGNORE_FILES = ['.DS_Store']
+const cwd = process.cwd()
+
+// Can be run as `yarn test:e2e --cache` to forego reinstalling node_modules, or
+// `yarn test:e2e <projects dir>`, or `yarn test:e2e --cache <projects dir>`.
+const args = process.argv.slice(2)
 
 function success(msg) {
   console.info(chalk.green('\n[vue-jest]: ' + msg + '\n'))
@@ -31,16 +36,30 @@ function runTest(dir) {
 
   const log = msg => info(`(${dir}) ${msg}`)
 
-  log('Running tests')
+  if (!args.filter(arg => arg === '--cache').length) {
+    log('Removing node_modules')
+    fs.removeSync(`${resolvedPath}/node_modules`)
 
-  log('Removing node_modules')
-  fs.removeSync(`${resolvedPath}/node_modules`)
+    log('Removing yarn-lock.json')
+    fs.removeSync(`${resolvedPath}/yarn-lock.json`)
 
-  log('Removing yarn-lock.json')
-  fs.removeSync(`${resolvedPath}/yarn-lock.json`)
+    log('Installing node_modules')
+    run('yarn install --silent')
+  }
 
-  log('Installing node_modules')
-  run('yarn install --silent')
+  // For tests that need vue-jest to successfully `require.resolve()` a file in
+  // the project directory's node_modules, we can't symlink vue-jest from a
+  // parent directory (as node module resolution walks up the file tree,
+  // starting from the realpath of the caller), we must copy it.
+  if (
+    !fs.existsSync(`${resolvedPath}/node_modules/vue-jest`) ||
+    !fs.lstatSync(`${resolvedPath}/node_modules/vue-jest`).isSymbolicLink()
+  ) {
+    log('Copying vue-jest into node_modules')
+    fs.mkdirSync(`${resolvedPath}/node_modules/vue-jest`, { recursive: true })
+    run(`cp ${cwd}/package.json node_modules/vue-jest/`)
+    run(`cp -r ${cwd}/lib node_modules/vue-jest/`)
+  }
 
   log('Running tests')
   run('yarn test')
@@ -48,12 +67,18 @@ function runTest(dir) {
   success(`(${dir}) Complete`)
 }
 
-async function testRunner(dir) {
+async function testRunner() {
   const directories = fs
     .readdirSync(path.resolve(__dirname, '__projects__'))
     .filter(d => !IGNORE_FILES.includes(d))
 
-  directories.forEach(runTest)
+  const matches = args.filter(d => directories.includes(d))
+
+  if (matches.length) {
+    matches.forEach(runTest)
+  } else {
+    directories.forEach(runTest)
+  }
 }
 
 testRunner()
